Refactor: TotalSale에서 sale로 도메인 위치 변경 및 hour 파라미터 number로 변경
