+++ src/hooks/api/sales.ts
@@ -1,4 +1,10 @@
-import { getOneDaySales, getSales, getSalesParams } from "@services/sales";
+import {
+  getOneDaySales,
+  getSales,
+  getSalesParams,
+  getTotalSales,
+  getTotalSalesParams,
+} from "@services/sales";
 import { useQuery } from "@tanstack/react-query";
 
 export const useGetOneDaySales = (date: string) => {
@@ -12,5 +18,12 @@ export const useGetSales = (params: getSalesParams) => {
   return useQuery({
     queryKey: ["sales"],
     queryFn: () => getSales(params),
+  });
+};
+
+export const useGetTotalSales = (params: getTotalSalesParams) => {
+  return useQuery({
+    queryKey: ["totalSales"],
+    queryFn: () => getTotalSales(params),
   });
 };
