+++ src/services/sales.ts
@@ -1,11 +1,12 @@
+import { TotalSales } from "@/models/totalSales.model";
 import { getRequest } from "./api";
 import { DailySales } from "@/models/dailySales.model";
 
 export type getSalesParams = {
   startDate: string;
   endDate: string;
-  startHour?: string;
-  endHour?: string;
+  startHour?: number;
+  endHour?: number;
 };
 
 export const getSales = async (params: getSalesParams) => {
@@ -33,4 +34,23 @@ export const getOneDaySales = async (params: getDailySalesParams) => {
     }
   );
   return response[0];
+};
+
+export type getTotalSalesParams = {
+  startDate: string;
+  endDate: string;
+  startHour?: number;
+  endHour?: number;
+};
+
+export const getTotalSales = async (params: getTotalSalesParams) => {
+  const response = await getRequest<TotalSales>(
+    "/mocks/totalSalesData.json", //API URL: /sales/total/{id}
+    {
+      ...params,
+      startDate: new Date(params.startDate).toISOString(),
+      endDate: new Date(params.endDate).toISOString(),
+    }
+  );
+  return response;
 };
